Move TokenType to Tokenizer
