+++ src/controllers/filters/parser/Tokenizer.cpp
@@ -4,6 +4,78 @@
 
 namespace filterparser {
 
+QString tokenTypeToInfoString(TokenType type)
+{
+    switch (type)
+    {
+        case CONTROL_START:
+        case CONTROL_END:
+        case BINARY_START:
+        case BINARY_END:
+        case UNARY_START:
+        case UNARY_END:
+        case MATH_START:
+        case MATH_END:
+        case OTHER_START:
+        case NONE:
+            return "<unknown>";
+        case AND:
+            return "<and>";
+        case OR:
+            return "<or>";
+        case LP:
+            return "<left parenthesis>";
+        case RP:
+            return "<right parenthesis>";
+        case LIST_START:
+            return "<list start>";
+        case LIST_END:
+            return "<list end>";
+        case COMMA:
+            return "<comma>";
+        case PLUS:
+            return "<plus>";
+        case MINUS:
+            return "<minus>";
+        case MULTIPLY:
+            return "<multiply>";
+        case DIVIDE:
+            return "<divide>";
+        case MOD:
+            return "<modulus>";
+        case EQ:
+            return "<equals>";
+        case NEQ:
+            return "<not equals>";
+        case LT:
+            return "<less than>";
+        case GT:
+            return "<greater than>";
+        case LTE:
+            return "<less than equal>";
+        case GTE:
+            return "<greater than equal>";
+        case CONTAINS:
+            return "<contains>";
+        case STARTS_WITH:
+            return "<starts with>";
+        case ENDS_WITH:
+            return "<ends with>";
+        case MATCH:
+            return "<match>";
+        case NOT:
+            return "<not>";
+        case STRING:
+            return "<string>";
+        case INT:
+            return "<int>";
+        case IDENTIFIER:
+            return "<identifier>";
+        default:
+            return "<unknown>";
+    }
+}
+
 Tokenizer::Tokenizer(const QString &text)
 {
     QRegularExpressionMatchIterator i = tokenRegex.globalMatch(text);
